Flyttet ut datepicker i to kompoonenter
