+++ apps/frontend/src/components/Process/DateFieldsProcessModal.tsx
@@ -1,11 +1,10 @@
 import { faQuestionCircle } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { Button, DatePicker, Tooltip } from '@navikt/ds-react'
-import { Field, FieldProps } from 'formik'
+import { Button, Tooltip } from '@navikt/ds-react'
 import { useState } from 'react'
-import { IProcessFormValues } from '../../constants'
 import { theme } from '../../util'
-import { Error } from '../common/ModalSchema'
+import { EndDate } from './EndDate'
+import { StartDate } from './StartDate'
 
 interface IDateModalProps {
   showDates: boolean
@@ -70,58 +69,9 @@ export const DateFieldsProcessModal = (props: IDateModalProps) => {
               </div>
             </div>
             <div className="flex w-full">
-              <div className="w-1/2 mr-4">
-                <div className="flex w-full mt-4">
-                  <Field name="start">
-                    {({ form }: FieldProps<string, IProcessFormValues>) => (
-                      <DatePicker
-                        onSelect={(date: any) => {
-                          const dateSingle: Date = Array.isArray(date) ? date[0] : date
-                          if (dateSingle) {
-                            const newDate = dateSingle.setDate(dateSingle.getDate() + 1)
-                            const formatedDate = new Date(newDate)
-                            form.setFieldValue('start', formatedDate.toISOString().split('T')[0])
-                          } else form.setFieldValue('start', undefined)
-                        }}
-                      >
-                        <DatePicker.Input
-                          className="mb-2"
-                          value={form.values['start']}
-                          label="Velg fra og med dato"
-                          error={!!form.errors.start && (form.touched.start || !!form.submitCount)}
-                        />
-                      </DatePicker>
-                    )}
-                  </Field>
-                </div>
-                <Error fieldName="start" />
-              </div>
-              <div className="w-1/2">
-                <div className="flex w-full mt-4">
-                  <Field name="end">
-                    {({ form }: FieldProps<string, IProcessFormValues>) => (
-                      <DatePicker
-                        onSelect={(date: any) => {
-                          const dateSingle: Date = Array.isArray(date) ? date[0] : date
-                          if (dateSingle) {
-                            const newDate = dateSingle.setDate(dateSingle.getDate() + 1)
-                            const formatedDate = new Date(newDate)
-                            form.setFieldValue('end', formatedDate.toISOString().split('T')[0])
-                          } else form.setFieldValue('end', undefined)
-                        }}
-                      >
-                        <DatePicker.Input
-                          className="mb-2"
-                          value={form.values['end']}
-                          label="Velg til og med dato"
-                          error={!!form.errors.end && (form.touched.end || !!form.submitCount)}
-                        />
-                      </DatePicker>
-                    )}
-                  </Field>
-                </div>
-                <Error fieldName="end" />
-              </div>
+              <StartDate />
+
+              <EndDate />
             </div>
           </div>
         </>
